pyproject.toml
[project]
name = "job-finder"
version = "0.1.0"
description = "AI-powered job search agent using CrewAI + Claude for data engineering leadership roles"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.95.0,<1.0.0",
    "python-jobspy>=1.1.75",
    "streamlit>=1.40.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
    "PyPDF2>=3.0.0",
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "litellm>=1.50.0",
]

[project.scripts]
job-finder = "job_finder.main:run"
search-jobs = "job_finder.main:search"
score-jobs = "job_finder.main:score"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_finder"]
